ENH : Simplify Chatbot Routing Structure
- Renamed chat sub-route to base id-based route for individual chat instances
- Removed redundant page component in chat sub-route
- Updated routes on Drawer component to match new structure
- Refactored route updating in useChat hook to follow new structure
- Integrated Chat component and necessary hooks directly into main page

This commit was generated by lguibr/comai
